+++ lxd/instance/drivers/qmp/monitor.go
